# api/gateway/analysis/api.yml
...
swagger: '2.0'
info:
  version: '1.0.0' # API version, not the SAR code version
  title: panther-analysis-api
  description: Panther Python policies/rules
  contact:
    name: Panther Labs
    url: https://runpanther.io/about
    email: support@runpanther.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

# The name of the Lambda handler function that will process API requests
...
parameters: # Parameters shared across multiple methods
...
  dataModelId:
    name: dataModelId
    in: query
    description: Unique ASCII dataModel identifier
    required: true
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
...
  versionId:
    name: versionId
    in: query
    description: The version of the analysis to retrieve
    type: string
    pattern: '[a-zA-Z\._0-9]{32}'
paths:
  /datamodel:
    # The UI datamodel detail shows all details for an individual user provided data model.
    #
    # Example: GET /datamodel? dataModelId=onelogin
    #
    # Response: {
    #     "body":                      "def mapping_method(event): return event.get('a field')",
    #     "createdAt":                 "2019-08-26T00:00:00.000Z",
    #     "createdBy":                 "5f54cf4a-ec56-44c2-83bc-8b742600f307",
    #     "description":               "This policy is always compliant",
    #     "enabled":                   true,
    #     "id":                        "OneLogin.DataModel",
    #     "lastModified":              "2019-08-26T00:00:00.000Z",
    #     "lastModifiedBy":            "5f54cf4a-ec56-44c2-83bc-8b742600f307",
    #     "logTypes":                  ["OneLogin.Events"],
    #     "mappings":                  [{"field":"ipAddr","name":"source_address"}],
    #     "versionId": "TsKejJ6GGi_KdH65g2iu9bcww8JxkkwI"
    # }
    get:
      operationId: GetDataModel
      summary: Get data model details
      parameters:
        - $ref: '#/parameters/dataModelId'
        - $ref: '#/parameters/versionId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/DataModel'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        404:
          description: DataModel does not exist
        500:
          description: Internal server error

    # When users create a new dataModel, the frontend uses this endpoint.
    #
    # This is almost identical to updating a dataModel, except the dataModelId must not already exist.
    #
    # Example: (see ModifyDataModel)
    post:
      operationId: CreateDataModel
      summary: Create a new Data Model module
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateDataModel'
      responses:
        201:
          description: DataModel created successfully
          schema:
            $ref: '#/definitions/DataModel'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        409:
          description: DataModel with the given ID already exists
        500:
          description: Internal server error
...
  /datamodel/update:
    # Same as UpdatePolicy, but for a dataModel module
    post:
      operationId: ModifyDataModel
      summary: Modify an existing Data Model
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateDataModel'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/DataModel'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        404:
          description: DataModel not found
        500:
          description: Internal server error
...
definitions:
  Error:
    type: object
    properties:
      message:
        description: Error message
        type: string
    required:
      - message

  ##### Create/Update/Modify #####
...
  TypeSet:
    description: List of resource/log types to which this policy applies
    type: array
    maxItems: 500
    uniqueItems: true
    items:
      type: string
      minLength: 1
...
  DataModel:
    type: object
    properties:
      body:
        $ref: '#/definitions/body'
      createdAt:
        $ref: '#/definitions/modifyTime'
      createdBy:
        $ref: '#/definitions/userId'
      description:
        $ref: '#/definitions/description'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      lastModified:
        $ref: '#/definitions/modifyTime'
      lastModifiedBy:
        $ref: '#/definitions/userId'
      logTypes:
        $ref: '#/definitions/TypeSet'
      mappings:
        $ref: '#/definitions/Mappings'
      versionId:
        $ref: '#/definitions/versionId'
    required:
      # Note: body is optional for DataModel
      - createdAt
      - createdBy
      - enabled
      - id
      - lastModified
      - lastModifiedBy
      - logTypes
      - mappings
      - versionId
  UpdateDataModel:
    type: object
    properties:
      body:
        $ref: '#/definitions/body'
      description:
        $ref: '#/definitions/description'
      enabled:
        $ref: '#/definitions/enabled'
      id:
        $ref: '#/definitions/id'
      logTypes:
        $ref: '#/definitions/TypeSet'
      mappings:
        $ref: '#/definitions/Mappings'
      userId:
        $ref: '#/definitions/userId'
    required:
      - enabled
      - id
      - mappings
      - userId

  ##### Create/Modify/Update Globals #####
...
  Mapping:
    type: object
    properties:
      name:
        $ref: '#/definitions/sourceName'
      field:
        $ref: '#/definitions/field'
      method:
        $ref: '#/definitions/method'
    required:
      - name
    #oneOf:
    #  - required: [method, name]
    #  - required: [field, name]
  Mappings:
    description: yml schema mappings
    type: array
    items:
      $ref: '#/definitions/Mapping'
...
  body:
    description: Python policy source code
    type: string
    minLength: 10
    maxLength: 1000000 # ~1 MB
...
  description:
    description: Summary of the policy and its purpose
    type: string
    maxLength: 5000
...
  enabled:
    description: True if the policy is currently being evaluated
    type: boolean
  field:
    description: field name used to map data in a DataModel
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
  id:
    description: User-specified unique rule/policy ID
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
  method:
    description: method name used to map data in a DataModel
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
  modifyTime:
    description: Policy modification timestamp
    type: string
    format: date-time
  sourceName:
    description: standardized field name used in DataModel
    type: string
    pattern: '[a-zA-Z0-9\-\. ]{1,200}'
...
  userId:
    description: Panther user ID that created or modified the policy
    type: string
    pattern: '[a-f0-9\-]{36}'
  versionId:
    description: Policy version ID
    type: string
    pattern: '[a-zA-Z\._0-9]{32}' # 32-character alphanumeric S3 object version